streamlit_app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes -o streamlit_app/requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
boto3==1.40.41
    # via sake-sensei
botocore==1.40.41
    # via
    #   boto3
    #   s3transfer
    #   sake-sensei
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==46.0.1
    # via pyjwt
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via sake-sensei
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via sake-sensei
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.6.0
    # via altair
numpy==2.3.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
orjson==3.11.3
    # via sake-sensei
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.1
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.11.9
    # via
    #   pydantic-settings
    #   sake-sensei
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via sake-sensei
pydeck==0.9.1
    # via streamlit
pyjwt==2.10.1
    # via sake-sensei
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
    #   sake-sensei
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   sake-sensei
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.14.0
    # via boto3
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
streamlit==1.50.0
    # via sake-sensei
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via streamlit
